+++ StrikeNeck/DB/hours.sql
@@ -28,3 +28,146 @@ INSERT INTO Test VALUES(20,0,0)
 INSERT INTO Test VALUES(21,0,0)
 INSERT INTO Test VALUES(22,0,0)
 INSERT INTO Test VALUES(23,0,0)
+INSERT INTO Test VALUES(24,0,0)
+INSERT INTO Test VALUES(25,0,0)
+INSERT INTO Test VALUES(26,0,0)
+INSERT INTO Test VALUES(27,0,0)
+INSERT INTO Test VALUES(28,0,0)
+INSERT INTO Test VALUES(29,0,0)
+INSERT INTO Test VALUES(30,0,0)
+INSERT INTO Test VALUES(31,0,0)
+INSERT INTO Test VALUES(32,0,0)
+INSERT INTO Test VALUES(33,0,0)
+INSERT INTO Test VALUES(34,0,0)
+INSERT INTO Test VALUES(35,0,0)
+INSERT INTO Test VALUES(36,0,0)
+INSERT INTO Test VALUES(37,0,0)
+INSERT INTO Test VALUES(38,0,0)
+INSERT INTO Test VALUES(39,0,0)
+INSERT INTO Test VALUES(40,0,0)
+INSERT INTO Test VALUES(41,0,0)
+INSERT INTO Test VALUES(42,0,0)
+INSERT INTO Test VALUES(43,0,0)
+INSERT INTO Test VALUES(44,0,0)
+INSERT INTO Test VALUES(45,0,0)
+INSERT INTO Test VALUES(46,0,0)
+INSERT INTO Test VALUES(47,0,0)
+INSERT INTO Test VALUES(48,0,0)
+INSERT INTO Test VALUES(49,0,0)
+INSERT INTO Test VALUES(50,0,0)
+INSERT INTO Test VALUES(51,0,0)
+INSERT INTO Test VALUES(52,0,0)
+INSERT INTO Test VALUES(53,0,0)
+INSERT INTO Test VALUES(54,0,0)
+INSERT INTO Test VALUES(55,0,0)
+INSERT INTO Test VALUES(56,0,0)
+INSERT INTO Test VALUES(57,0,0)
+INSERT INTO Test VALUES(58,0,0)
+INSERT INTO Test VALUES(59,0,0)
+INSERT INTO Test VALUES(60,0,0)
+INSERT INTO Test VALUES(61,0,0)
+INSERT INTO Test VALUES(62,0,0)
+INSERT INTO Test VALUES(63,0,0)
+INSERT INTO Test VALUES(64,0,0)
+INSERT INTO Test VALUES(65,0,0)
+INSERT INTO Test VALUES(66,0,0)
+INSERT INTO Test VALUES(67,0,0)
+INSERT INTO Test VALUES(68,0,0)
+INSERT INTO Test VALUES(69,0,0)
+INSERT INTO Test VALUES(70,0,0)
+INSERT INTO Test VALUES(71,0,0)
+INSERT INTO Test VALUES(72,0,0)
+INSERT INTO Test VALUES(73,0,0)
+INSERT INTO Test VALUES(74,0,0)
+INSERT INTO Test VALUES(75,0,0)
+INSERT INTO Test VALUES(76,0,0)
+INSERT INTO Test VALUES(77,0,0)
+INSERT INTO Test VALUES(78,0,0)
+INSERT INTO Test VALUES(79,0,0)
+INSERT INTO Test VALUES(80,0,0)
+INSERT INTO Test VALUES(81,0,0)
+INSERT INTO Test VALUES(82,0,0)
+INSERT INTO Test VALUES(83,0,0)
+INSERT INTO Test VALUES(84,0,0)
+INSERT INTO Test VALUES(85,0,0)
+INSERT INTO Test VALUES(86,0,0)
+INSERT INTO Test VALUES(87,0,0)
+INSERT INTO Test VALUES(88,0,0)
+INSERT INTO Test VALUES(89,0,0)
+INSERT INTO Test VALUES(90,0,0)
+INSERT INTO Test VALUES(91,0,0)
+INSERT INTO Test VALUES(92,0,0)
+INSERT INTO Test VALUES(93,0,0)
+INSERT INTO Test VALUES(94,0,0)
+INSERT INTO Test VALUES(95,0,0)
+INSERT INTO Test VALUES(96,0,0)
+INSERT INTO Test VALUES(97,0,0)
+INSERT INTO Test VALUES(98,0,0)
+INSERT INTO Test VALUES(99,0,0)
+INSERT INTO Test VALUES(100,0,0)
+INSERT INTO Test VALUES(101,0,0)
+INSERT INTO Test VALUES(102,0,0)
+INSERT INTO Test VALUES(103,0,0)
+INSERT INTO Test VALUES(104,0,0)
+INSERT INTO Test VALUES(105,0,0)
+INSERT INTO Test VALUES(106,0,0)
+INSERT INTO Test VALUES(107,0,0)
+INSERT INTO Test VALUES(108,0,0)
+INSERT INTO Test VALUES(109,0,0)
+INSERT INTO Test VALUES(110,0,0)
+INSERT INTO Test VALUES(111,0,0)
+INSERT INTO Test VALUES(112,0,0)
+INSERT INTO Test VALUES(113,0,0)
+INSERT INTO Test VALUES(114,0,0)
+INSERT INTO Test VALUES(115,0,0)
+INSERT INTO Test VALUES(116,0,0)
+INSERT INTO Test VALUES(117,0,0)
+INSERT INTO Test VALUES(118,0,0)
+INSERT INTO Test VALUES(119,0,0)
+INSERT INTO Test VALUES(120,0,0)
+INSERT INTO Test VALUES(121,0,0)
+INSERT INTO Test VALUES(122,0,0)
+INSERT INTO Test VALUES(123,0,0)
+INSERT INTO Test VALUES(124,0,0)
+INSERT INTO Test VALUES(125,0,0)
+INSERT INTO Test VALUES(126,0,0)
+INSERT INTO Test VALUES(127,0,0)
+INSERT INTO Test VALUES(128,0,0)
+INSERT INTO Test VALUES(129,0,0)
+INSERT INTO Test VALUES(130,0,0)
+INSERT INTO Test VALUES(131,0,0)
+INSERT INTO Test VALUES(132,0,0)
+INSERT INTO Test VALUES(133,0,0)
+INSERT INTO Test VALUES(134,0,0)
+INSERT INTO Test VALUES(135,0,0)
+INSERT INTO Test VALUES(136,0,0)
+INSERT INTO Test VALUES(137,0,0)
+INSERT INTO Test VALUES(138,0,0)
+INSERT INTO Test VALUES(139,0,0)
+INSERT INTO Test VALUES(140,0,0)
+INSERT INTO Test VALUES(141,0,0)
+INSERT INTO Test VALUES(142,0,0)
+INSERT INTO Test VALUES(143,0,0)
+INSERT INTO Test VALUES(144,0,0)
+INSERT INTO Test VALUES(145,0,0)
+INSERT INTO Test VALUES(146,0,0)
+INSERT INTO Test VALUES(147,0,0)
+INSERT INTO Test VALUES(148,0,0)
+INSERT INTO Test VALUES(149,0,0)
+INSERT INTO Test VALUES(150,0,0)
+INSERT INTO Test VALUES(151,0,0)
+INSERT INTO Test VALUES(152,0,0)
+INSERT INTO Test VALUES(153,0,0)
+INSERT INTO Test VALUES(154,0,0)
+INSERT INTO Test VALUES(155,0,0)
+INSERT INTO Test VALUES(156,0,0)
+INSERT INTO Test VALUES(157,0,0)
+INSERT INTO Test VALUES(158,0,0)
+INSERT INTO Test VALUES(159,0,0)
+INSERT INTO Test VALUES(160,0,0)
+INSERT INTO Test VALUES(161,0,0)
+INSERT INTO Test VALUES(162,0,0)
+INSERT INTO Test VALUES(163,0,0)
+INSERT INTO Test VALUES(164,0,0)
+INSERT INTO Test VALUES(165,0,0)
+INSERT INTO Test VALUES(166,0,0)
